+++ src/main/java/com/macuguita/lib/reg/GuitaRegistryEntries.java
@@ -9,8 +9,8 @@ public class GuitaRegistryEntries<T> {
 
     private final List<GuitaRegistryEntry<T>> entries = new ArrayList<>();
 
+    @SuppressWarnings("unchecked")
     public <I extends T, E extends GuitaRegistryEntry<I>> E add(E entry) {
-        //noinspection unchecked
         entries.add((GuitaRegistryEntry<T>) entry);
         return entry;
     }