+++ src/testmod/java/com/macuguita/libtest/neoforge/NeoforgeEntrypoint.java
@@ -1,0 +1,14 @@
+package com.macuguita.libtest.neoforge;
+
+//? neoforge {
+/*import com.macuguita.libtest.TestMod;
+import net.neoforged.fml.common.Mod;
+
+@Mod(TestMod.MOD_ID)
+public class NeoforgeEntrypoint {
+
+    public NeoforgeEntrypoint() {
+        TestMod.init();
+    }
+}
+*///?}