+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 2.0.0-SNAPSHOT
-- # This is a test build
+- Made the supporter cape system better
 - Update to latest minecraft version
 - Simplified the api
 ---